ref(ngrx-ducks): move mocks and new createDuck to own directories
