+++ ngrx/ducks/test/mocks/duck-candidates/index.ts
