+++ ngrx/ducks/test/mocks/duck-candidates/with-property.ts
