+++ ngrx/ducks/test/mocks/duck-candidates/without-action-decorator.ts
