+++ ngrx/ducks/test/mocks/index.ts
@@ -1,1 +1,2 @@
+export * from './duck-candidates';
 export * from './store.mock';
